refactor(): rename martial-arts files with their folder names
